Improved Trait registry
You can register by ID or by class now, and if the ID doesn't exist it gets ignored. That way, in the future, either myself or mods can add traits to blocks that might not be present in that particular modpack.

I used the class registry to apply the OxidizableTrait, and an ID registry for `blockus:redstone_brick_slab`