Vertical Slab Tool Category
Adds vertical slabs to the same `mineable` tag as their parent. Does this by checking every tag that has the word `mineable` in it for slabs and adding the vertical slabs to it, so it's none of the other tags. This should also snag some modded tool tags, we'll see