What if we changed everything
Alright, so this was a big sweep of changes. After all that slab context work, I removed the concept entirely in favor of an interface called SlabLike that lets you do slab operations. This also removed all the forwarding functions built into MixedSlabBlock, and moved most of the Utility functions out of that class into SlabLike. Why I didn't do this from the start, the world may never know.

Also of note, fixed a compat bug with Mo Glass