That wasn't too bad
Had to add a blockstate to mixed slabs though, rather unfortunate. I was able to use dynamic emission on NeoForge if I wanted, but Fabric doesn't have another solution without messing with the lighting system a lot more than a slab mod should, so I did the same on both for ease.